requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
altgraph @ file:///System/Volumes/Data/SWE/Apps/DT/BuildRoots/BuildRoot2/ActiveBuildRoot/Library/Caches/com.apple.xbs/Sources/python3/python3-124/altgraph-0.17.2-py2.py3-none-any.whl
asgiref==3.6.0
async-timeout==4.0.2
attrs==22.2.0
beautifulsoup4==4.11.2
certifi==2022.12.7
charset-normalizer==2.1.1
Django==4.1.6
django-cors-headers==3.13.0
django-endless-pagination==2.0
django-livereload-server==0.4
djangorestframework==3.14.0
frozenlist==1.3.3
future @ file:///System/Volumes/Data/SWE/Apps/DT/BuildRoots/BuildRoot2/ActiveBuildRoot/Library/Caches/com.apple.xbs/Sources/python3/python3-124/future-0.18.2-py3-none-any.whl
idna==3.4
livereload==2.6.3
macholib @ file:///System/Volumes/Data/SWE/Apps/DT/BuildRoots/BuildRoot2/ActiveBuildRoot/Library/Caches/com.apple.xbs/Sources/python3/python3-124/macholib-1.15.2-py2.py3-none-any.whl
multidict==6.0.4
mysqlclient==2.1.1
openai==0.26.4
Pillow==9.4.0
PyMySQL==1.0.2
pytz==2022.7.1
requests==2.28.2
# six @ file:///System/Volumes/Data/SWE/Apps/DT/BuildRoots/BuildRoot2/ActiveBuildRoot/Library/Caches/com.apple.xbs/Sources/python3/python3-124/six-1.15.0-py2.py3-none-any.whl
soupsieve==2.3.2.post1
sqlparse==0.4.3
tornado==6.2
tqdm==4.64.1
urllib3==1.26.14
whitenoise==6.3.0
yarl==1.8.2

fuzzywuzzy~=0.18.0
djoser~=2.1.0
six~=1.15.0